pyproject.toml
[tool.poetry]
name = "mesh3d-generator"
version = "0.1.0"
description = "Geração de malhas 3D a partir de imagens e texto descritivo usando LLM"
authors = ["Seu Nome <seu.email@example.com>"]
readme = "README.md"
packages = [{include = "mesh3d_generator"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.9.0"
torchvision = "^0.24.0"
transformers = "^4.57.0"
diffusers = "^0.35.0"
accelerate = "^0.25.0"
pillow = "^10.1.0"
numpy = "^1.26.0"
opencv-python = "^4.9.0"
trimesh = "^3.23.0"
# open3d instalado manualmente: poetry run pip install open3d
scipy = "^1.11.0"
matplotlib = "^3.8.0"
tqdm = "^4.66.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.0"
requests = "^2.31.0"
gdown = "^4.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true


